Seperated bed class from RoomAbstract, helped with bugs
