+++ PrimeTimeApp/src/main/java/org/primeTimeHotel/Controllers/RoomController.java
@@ -2,7 +2,6 @@ package org.primeTimeHotel.Controllers;
 
 
 import org.primeTimeHotel.Database_Objects.RoomDAO;
-import org.primeTimeHotel.Domain_Model_Objects.RoomAbstractClass;
 
 public class RoomController extends RoomDAO {
