+++ PrimeTimeApp/src/main/java/org/primeTimeHotel/Main.java
@@ -5,6 +5,7 @@ import org.primeTimeHotel.Database_Objects.MasterDAO;
 import org.primeTimeHotel.Database_Objects.ReservationDAO;
 import org.primeTimeHotel.Database_Objects.RoomDAO;
 import org.primeTimeHotel.Domain_Model_Objects.*;
+import org.primeTimeHotel.Domain_Model_Objects.RoomClasses.Bed;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -12,7 +13,6 @@ import java.util.List;
 public class Main {
     public static void main(String[] args) {
         RoomDAO dao = new RoomDAO();
-
